Move custom-element code to lite/custom-element module
